Refactor AST. Add Semantic check file
