+++ compile.sh
@@ -6,13 +6,15 @@ flex lexicon.l
 
 echo "Compiling sources"
 gcc -Wall -c src/ast.c -o build/ast.o
+gcc -Wall -c src/ast_modules.c -o build/ast_modules.o
 gcc -Wall -c src/symbols.c -o build/symbols.o
+gcc -Wall -c src/semantic_check.c -o build/semantic_check.o
 gcc -Wall -c src/utils.c -o build/utils.o
 gcc -Wall -c src/main.c -o build/main.o
 
 echo "Linking"
 gcc -Wall -Wno-unused-function syntax.tab.c lex.yy.c \
-  build/main.o build/utils.o build/ast.o build/symbols.o \
+  build/main.o build/utils.o build/ast.o build/symbols.o build/ast_modules.o build/semantic_check.o \
   -o build/c-tds -lfl
 
 # Create symbolic link with the output name
